requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bcrypt==4.2.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
coverage==7.5.3
cryptography==43.0.1
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
django-resized==1.0.2
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==27.0.0
google-api-core==2.19.1
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
packaging==24.1
paramiko==3.4.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.2.2
ruff==0.6.9